--- dev/LayoutExploerer.module.css
@@ -1,57 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.logo {
-  animation: logo-spin infinite 20s linear;
-  height: 40vmin;
-  pointer-events: none;
-}
-
-.header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.link {
-  color: #b318f0;
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.block {
-  height: 150px;
-  background-color: red;
-  line-height: 100%;
-  text-align: center;
-  font-size: 50px;
-  user-select: none;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-  align-items: start;
-}
-
-.handle {
-  width: 50px;
-  height: 50px;
-  background-color: blue;
-}
-
-.button {
-  margin: 24px;
-  padding: 8px;
-}
